wasapi: Patched WinRT to compile.
